HS-361: BTO: Event Service - Add endpoint to get Events by node id

+++ rest-server/src/main/java/org/opennms/horizon/server/model/events/Event.java
@@ -13,7 +13,7 @@ public class Event {
     private String uei;
     private int nodeId;
     private String ipAddress;
-    private int producedTime;
+    private long producedTime;
     private List<EventParameter> eventParams;
     private EventInfo eventInfo;
 }
